fix: `XKitFeatureElement` extends `CustomElement`
