+++ src/action/popup.html
@@ -89,30 +89,6 @@
       </section>
     </main>
 
-    <template id="xkit-feature">
-      <link rel="stylesheet" href="../lib/normalize.min.css">
-      <link rel="stylesheet" href="../lib/remixicon/remixicon.css">
-      <link rel="stylesheet" href="../lib/spectrum.css">
-      <link rel="stylesheet" href="../lib/toggle-button.css">
-      <link rel="stylesheet" href="./components/xkit-feature/index.css">
-      <details>
-        <summary>
-          <div class="icon">
-            <slot name="icon"></slot>
-          </div>
-          <div class="meta">
-            <h4 class="title"><slot name="title"></slot></h4>
-            <p class="description"><slot name="description"></slot></p>
-          </div>
-          <div class="buttons">
-            <a class="help" target="_blank"><i class="ri-fw ri-question-fill" style="color:rgb(var(--black))"></i></a>
-            <input type="checkbox" checked class="toggle-button" aria-label="Enable this feature">
-          </div>
-        </summary>
-        <ul class="preferences"></ul>
-      </details>
-    </template>
-
     <template id="checkbox-preference">
       <li>
         <input type="checkbox">
